+++ database/migrations/2019_02_02_223332_init.php
@@ -32,15 +32,15 @@ class Init extends Migration
             $table->char('id', 26)->nullable(false);
             $table->char('user_id', 26)->nullable(false);
             $table->date('date')->nullable(false);
-            $table->integer('protein')->nullable(false);
-            $table->integer('fat')->nullable(false);
-            $table->integer('carbohydrates')->nullable(false);
-            $table->integer('fiber')->nullable(false);
+            $table->float('protein')->nullable(false);
+            $table->float('fat')->nullable(false);
+            $table->float('carbohydrates')->nullable(false);
+            $table->float('fiber')->nullable(false);
             $table->integer('weight')->nullable(false);
-            $table->integer('protein_eaten')->nullable(false);
-            $table->integer('fat_eaten')->nullable(false);
-            $table->integer('carbohydrates_eaten')->nullable(false);
-            $table->integer('fiber_eaten')->nullable(false);
+            $table->float('protein_eaten')->nullable(false);
+            $table->float('fat_eaten')->nullable(false);
+            $table->float('carbohydrates_eaten')->nullable(false);
+            $table->float('fiber_eaten')->nullable(false);
             $table->integer('weight_eaten')->nullable(false);
 
             $table->timestamps();
@@ -54,10 +54,10 @@ class Init extends Migration
             $table->char('day_id', 26)->nullable(false);
             $table->char('meal_id', 26)->nullable(false);
             $table->char('user_id', 26)->nullable(false);
-            $table->integer('protein')->nullable(false);
-            $table->integer('fat')->nullable(false);
-            $table->integer('carbohydrates')->nullable(false);
-            $table->integer('fiber')->nullable(false);
+            $table->float('protein')->nullable(false);
+            $table->float('fat')->nullable(false);
+            $table->float('carbohydrates')->nullable(false);
+            $table->float('fiber')->nullable(false);
             $table->integer('weight')->nullable(false);
             $table->boolean('eaten')->nullable(false);
             $table->time('time')->nullable();
